refactor(allure-cypress): factor out test runtime
